Write transform function (broken?).
Write the transform function. I think I have written the same function
that was in the blog post; however, when I run it, the code fails at
run-time because of an undefined value. However, if I print the value of
line, which I think might be undefined, everything works! Very
perplexing.